Added full webcam snapshot synchronization Added new signals:  + RequestTrackerImageSignal  + ReplyTrackerImageSignal  + BroadcastTrackerImageSignal  + ReceiveTrackerNumberSignal Fixed improper handling of StartSimulationSignal on client side Changed around tracker image updating on GUI code  + Local webcam image now stored separately Changed DisplayWindow GUI code so it is now properly scrollable and properly sized Added automatic settings saving on window close to server GUI Client now actually has a tracker number and waits until it has one before performing any tasks Added support for sending BufferedImages over the ByteStream  - JPEG only, no checksum support Clients with invalid (or not tracker number) are ignored on the server Signal output now indicates tracker number (if they have one) instead of client number Fixed inconsistencies where trackerNumber / trackerID was an int instead of a byte
Issues:
 + Random exceptions related to GUI painting occur occasionally, but don't interrupt anything
 + JPEGs are lossy and encoded differently on each end, so checksums are ignored

TODO's:
 + Fix inconsistencies between trackerID / trackerNumber (they're the same)

git-svn-id: file:///e/SVN/tsep@25 9e8b06f4-fe92-4974-b243-337f9da7b896

+++ MasterServer/src/server/TrackerIdentifier.java
@@ -95,11 +95,11 @@ public class TrackerIdentifier extends Thread {
 					continue;
 				}
 				
-				int trackerNumber = -1;
+				byte trackerNumber = -1;
 				for(int j=0;j<SettingsManager.defaultTrackerIPAddress.length;j++) {
 					InetAddress trackerIP = m_settings.getTrackerIPAddress(j + 1);
 					if(trackerIP.equals(c.getIPAddress())) {
-						trackerNumber = j + 1;
+						trackerNumber = (byte) (j + 1);
 						break;
 					}
 				}
@@ -123,7 +123,7 @@ public class TrackerIdentifier extends Thread {
 					} while(!validInput);
 					
 					try {
-						trackerNumber = Integer.parseInt(input.toString());
+						trackerNumber = Byte.parseByte(input.toString());
 						trackerIdentified = true;
 					}
 					catch(NumberFormatException e) { }
